adding laravel excel queue driver using redis, fixing laporan jalinan excel download, fixing kegiatan bkcu peserta table view

+++ app/Http/Controllers/AnggotaCuDraftController.php
@@ -72,18 +72,18 @@ class AnggotaCuDraftController extends Controller{
 			$kelas3 = AnggotaCu::create($data);
 
 			$data2 = array_map(function($dat) use ($kelas3) {
-					return array(
-							'anggota_cu_id' => $kelas3->id,
-							'cu_id' => $dat['cu_id'],
-							'tp_id' => $dat['tp_id'],
-							'no_ba' => $dat['no_ba'],
-							'tanggal_masuk' => $dat['tanggal_masuk'],
-							'tanggal_keluar' => $dat['tanggal_keluar'],
-							'keterangan_masuk' => $dat['keterangan_masuk'],
-							'keterangan_keluar' => $dat['keterangan_keluar'],
-							'created_at' => $dat['created_at'],
-							'updated_at' => $dat['updated_at'],
-					);
+				return array(
+					'anggota_cu_id' => $kelas3->id,
+					'cu_id' => $dat['cu_id'],
+					'tp_id' => $dat['tp_id'],
+					'no_ba' => $dat['no_ba'],
+					'tanggal_masuk' => $dat['tanggal_masuk'],
+					'tanggal_keluar' => $dat['tanggal_keluar'],
+					'keterangan_masuk' => $dat['keterangan_masuk'],
+					'keterangan_keluar' => $dat['keterangan_keluar'],
+					'created_at' => $dat['created_at'],
+					'updated_at' => $dat['updated_at'],
+				);
 			}, $data2);
 			
 			$kelas4 = AnggotaCuCu::insert($data2);
@@ -115,7 +115,7 @@ class AnggotaCuDraftController extends Controller{
 					$query->where('anggota_cu_cu_draft.cu_id',$cu);
 				});
 				// $kelas->nik = preg_replace('/[^A-Za-z0-9]/', '',$kelas->nik);
-			}
+		}
 
 			$datas = $kelas->get();
 			foreach($datas as $item){
@@ -131,24 +131,31 @@ class AnggotaCuDraftController extends Controller{
 
 				$data2 = array_map(function($dat) use ($kelas3) {
 					return array(
-							'anggota_cu_id' => $kelas3->id,
-							'cu_id' => $dat['cu_id'],
-							'tp_id' => $dat['tp_id'],
-							'no_ba' => $dat['no_ba'],
-							'tanggal_masuk' => $dat['tanggal_masuk'],
-							'tanggal_keluar' => $dat['tanggal_keluar'],
-							'keterangan_masuk' => $dat['keterangan_masuk'],
-							'keterangan_keluar' => $dat['keterangan_keluar'],
-							'created_at' => $dat['created_at'],
-							'updated_at' => $dat['updated_at'],
+						'anggota_cu_id' => $kelas3->id,
+						'cu_id' => $dat['cu_id'],
+						'tp_id' => $dat['tp_id'],
+						'no_ba' => $dat['no_ba'],
+						'tanggal_masuk' => $dat['tanggal_masuk'],
+						'tanggal_keluar' => $dat['tanggal_keluar'],
+						'keterangan_masuk' => $dat['keterangan_masuk'],
+						'keterangan_keluar' => $dat['keterangan_keluar'],
+						'created_at' => $dat['created_at'],
+						'updated_at' => $dat['updated_at'],
 					);
 				}, $data2);
 
 				$kelas4 = AnggotaCuCu::insert($data2);
+
+				if($cu != 'semua'){
+					AnggotaCuDraft::destroy($item->id);
+				}
+
 				$kelas2->delete();
 			}
 
-			$kelas->delete();
+			if($cu == 'semua'){
+				$kelas->delete();
+			}
 
 			\DB::commit();
 			
@@ -290,8 +297,8 @@ class AnggotaCuDraftController extends Controller{
 				foreach($datas as $item){
 					$kelas2 = AnggotaCuCuDraft::where('anggota_cu_draft_id', $item->id);
 					$kelas2->delete();
+					AnggotaCuDraft::destroy($item->id);
 				}
-				$kelas->delete();
 			}
 			
 			\DB::commit();
